[fix] fix comment
