+++ paimon-flink/paimon-flink-common/src/main/java/org/apache/paimon/flink/action/ActionBase.java
@@ -21,16 +21,24 @@ package org.apache.paimon.flink.action;
 import org.apache.paimon.catalog.Catalog;
 import org.apache.paimon.flink.FlinkCatalog;
 import org.apache.paimon.flink.FlinkCatalogFactory;
+import org.apache.paimon.flink.LogicalTypeConversion;
 import org.apache.paimon.flink.utils.StreamExecutionEnvironmentUtils;
 import org.apache.paimon.options.CatalogOptions;
 import org.apache.paimon.options.Options;
+import org.apache.paimon.types.DataType;
+import org.apache.paimon.types.DataTypeCasts;
 
 import org.apache.flink.configuration.PipelineOptions;
 import org.apache.flink.configuration.ReadableConfig;
 import org.apache.flink.streaming.api.environment.StreamExecutionEnvironment;
+import org.apache.flink.table.api.EnvironmentSettings;
+import org.apache.flink.table.api.bridge.java.StreamTableEnvironment;
+import org.apache.flink.table.types.logical.LogicalType;
 
+import java.util.List;
 import java.util.Map;
 import java.util.UUID;
+import java.util.stream.Collectors;
 
 /** Abstract base of {@link Action} for table. */
 public abstract class ActionBase implements Action {
@@ -40,6 +48,8 @@ public abstract class ActionBase implements Action {
     protected final Catalog catalog;
     protected final FlinkCatalog flinkCatalog;
     protected final String catalogName = "paimon-" + UUID.randomUUID();
+    protected final StreamExecutionEnvironment env;
+    protected final StreamTableEnvironment batchTEnv;
 
     public ActionBase(String warehouse, Map<String, String> catalogConfig) {
         catalogOptions = Options.fromMap(catalogConfig);
@@ -47,6 +57,12 @@ public abstract class ActionBase implements Action {
 
         catalog = FlinkCatalogFactory.createPaimonCatalog(catalogOptions);
         flinkCatalog = FlinkCatalogFactory.createCatalog(catalogName, catalog, catalogOptions);
+        env = StreamExecutionEnvironment.getExecutionEnvironment();
+        batchTEnv = StreamTableEnvironment.create(env, EnvironmentSettings.inBatchMode());
+
+        // register flink catalog to table environment
+        batchTEnv.registerCatalog(flinkCatalog.getName(), flinkCatalog);
+        batchTEnv.useCatalog(flinkCatalog.getName());
     }
 
     protected void execute(StreamExecutionEnvironment env, String defaultName) throws Exception {
@@ -59,5 +75,35 @@ public abstract class ActionBase implements Action {
         // to make the action workflow serializable
         Options catalogOptions = this.catalogOptions;
         return () -> FlinkCatalogFactory.createPaimonCatalog(catalogOptions);
+    }
+
+    /**
+     * Extract {@link LogicalType}s from Flink {@link org.apache.flink.table.types.DataType}s and
+     * convert to Paimon {@link DataType}s.
+     */
+    protected List<DataType> toPaimonTypes(
+            List<org.apache.flink.table.types.DataType> flinkDataTypes) {
+        return flinkDataTypes.stream()
+                .map(org.apache.flink.table.types.DataType::getLogicalType)
+                .map(LogicalTypeConversion::toDataType)
+                .collect(Collectors.toList());
+    }
+
+    /**
+     * Check whether each {@link DataType} of actualTypes is compatible with that of expectedTypes
+     * respectively.
+     */
+    protected boolean compatibleCheck(List<DataType> actualTypes, List<DataType> expectedTypes) {
+        if (actualTypes.size() != expectedTypes.size()) {
+            return false;
+        }
+
+        for (int i = 0; i < actualTypes.size(); i++) {
+            if (!DataTypeCasts.supportsCompatibleCast(actualTypes.get(i), expectedTypes.get(i))) {
+                return false;
+            }
+        }
+
+        return true;
     }
 }
