--- paimon-flink/paimon-flink-common/src/main/java/org/apache/paimon/flink/action/FlinkActionEnvironmentBase.java
@@ -1,86 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.paimon.flink.action;
-
-import org.apache.paimon.catalog.Identifier;
-import org.apache.paimon.flink.LogicalTypeConversion;
-import org.apache.paimon.types.DataType;
-import org.apache.paimon.types.DataTypeCasts;
-
-import org.apache.flink.streaming.api.environment.StreamExecutionEnvironment;
-import org.apache.flink.table.api.EnvironmentSettings;
-import org.apache.flink.table.api.bridge.java.StreamTableEnvironment;
-import org.apache.flink.table.types.logical.LogicalType;
-
-import java.util.List;
-import java.util.Map;
-import java.util.stream.Collectors;
-
-/** Flink Action Base with batch environment and stream environment. */
-public abstract class FlinkActionEnvironmentBase extends ActionBase {
-
-    protected final StreamExecutionEnvironment env;
-    protected final StreamTableEnvironment batchTEnv;
-    protected final Identifier identifier;
-
-    FlinkActionEnvironmentBase(
-            String warehouse,
-            String databaseName,
-            String tableName,
-            Map<String, String> catalogConfig) {
-        super(warehouse, catalogConfig);
-        env = StreamExecutionEnvironment.getExecutionEnvironment();
-        batchTEnv = StreamTableEnvironment.create(env, EnvironmentSettings.inBatchMode());
-
-        // register flink catalog to table environment
-        batchTEnv.registerCatalog(flinkCatalog.getName(), flinkCatalog);
-        batchTEnv.useCatalog(flinkCatalog.getName());
-        identifier = new Identifier(databaseName, tableName);
-    }
-
-    /**
-     * Extract {@link LogicalType}s from Flink {@link org.apache.flink.table.types.DataType}s and
-     * convert to Paimon {@link DataType}s.
-     */
-    protected List<DataType> toPaimonTypes(
-            List<org.apache.flink.table.types.DataType> flinkDataTypes) {
-        return flinkDataTypes.stream()
-                .map(org.apache.flink.table.types.DataType::getLogicalType)
-                .map(LogicalTypeConversion::toDataType)
-                .collect(Collectors.toList());
-    }
-
-    /**
-     * Check whether each {@link DataType} of actualTypes is compatible with that of expectedTypes
-     * respectively.
-     */
-    protected boolean compatibleCheck(List<DataType> actualTypes, List<DataType> expectedTypes) {
-        if (actualTypes.size() != expectedTypes.size()) {
-            return false;
-        }
-
-        for (int i = 0; i < actualTypes.size(); i++) {
-            if (!DataTypeCasts.supportsCompatibleCast(actualTypes.get(i), expectedTypes.get(i))) {
-                return false;
-            }
-        }
-
-        return true;
-    }
-}
